requirements.txt
pytube
python-dotenv
discord.py
PyNaCl